refactor: components were moved to the separate directories
