+++ src/App.tsx
@@ -3,9 +3,9 @@ import { Component } from 'react';
 import type { Character } from './types/types.ts';
 
 import { fetchCharacters } from './api/characters-api';
-import { Header } from './components/Header';
-import { Loader } from './components/Loader';
-import { Main } from './components/Main';
+import { Header } from './components/Header/Header.tsx';
+import { Loader } from './components/Loader/Loader.tsx';
+import { Main } from './components/Main/Main.tsx';
 
 type AppProps = object;
