+++ src/index.tsx
@@ -3,8 +3,8 @@ import { createRoot } from 'react-dom/client';
 
 import './index.css';
 import { App } from './App.tsx';
-import { ErrorBoundary } from './components/ErrorBoundary.tsx';
-import { Fallback } from './components/Fallback.tsx';
+import { ErrorBoundary } from './components/ErrorBoundary/ErrorBoundary.tsx';
+import { Fallback } from './components/Fallback/Fallback.tsx';
 
 const rootElement = document.getElementById('root');
